Avoid dependency on get_distribution
This is only useful if we use setuptools to manage our deps, i.e.,
install pymediainfo into some system wide path.

For simplicity, we import from source code directly, and that source
is controlled under citw github directly, as a git submodule.

The benefit is so that user don't need to install anything, the repo
should work out of box once clonely properly (use recursive clone to
get the submodules).